LayoutTests/fast/images/text-recognition/ios/select-word-in-image-overlay.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../resources/js-test.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
<style>
img {
    position: absolute;
    top: 0;
    left: 0;
}
</style>
</head>
<body>
<img src="../../resources/green-400x400.png"></img>
<script>
jsTestIsAsync = true;
addEventListener("load", async () => {
    let image = document.querySelector("img");
    internals.installImageOverlay(image, [
        {
            topLeft : new DOMPointReadOnly(0, 0),
            topRight : new DOMPointReadOnly(1, 0),
            bottomRight : new DOMPointReadOnly(1, 0.5),
            bottomLeft : new DOMPointReadOnly(0, 0.5),
            children: [
                {
                    text : "hello",
                    topLeft : new DOMPointReadOnly(0, 0),
                    topRight : new DOMPointReadOnly(1, 0),
                    bottomRight : new DOMPointReadOnly(1, 0.5),
                    bottomLeft : new DOMPointReadOnly(0, 0.5),
                }
            ],
        },
        {
            topLeft : new DOMPointReadOnly(0, 0.5),
            topRight : new DOMPointReadOnly(1, 0.5),
            bottomRight : new DOMPointReadOnly(1, 1),
            bottomLeft : new DOMPointReadOnly(0, 1),
            children: [
                {
                    text : "world",
                    topLeft : new DOMPointReadOnly(0, 0.5),
                    topRight : new DOMPointReadOnly(1, 0.5),
                    bottomRight : new DOMPointReadOnly(1, 1),
                    bottomLeft : new DOMPointReadOnly(0, 1),
                }
            ],
        }
    ]);

    await UIHelper.longPressAtPoint(200, 300);
    selectionRects = await UIHelper.waitForSelectionToAppear();
    shouldBe("selectionRects.length", "1");
    finishJSTest();
});
</script>
</body>
</html>
